[refactor] Extract agents from module folders
