+++ src/manager/agents/init-module/module/init_module.cpp
@@ -6,7 +6,7 @@
 
 #include "init_module.hpp"
 
-#include "agents/sc_component_manager_agent_init.hpp"
+#include "agent/sc_component_manager_agent_init.hpp"
 
 using namespace initModule;
