+++ src/manager/agents/init-module/test/units/agent_tests.cpp
@@ -8,7 +8,7 @@
 
 #include <common/sc_component_manager_keynodes.hpp>
 
-#include "agents/sc_component_manager_agent_init.hpp"
+#include "agent/sc_component_manager_agent_init.hpp"
 
 TEST_F(ScAgentsTest, AgentInit)
 {
