+++ src/manager/agents/search-module/module/search-module.cpp
@@ -6,7 +6,7 @@
 
 #include "search-module.hpp"
 
-#include "agents/sc_component_manager_agent_search.hpp"
+#include "agent/sc_component_manager_agent_search.hpp"
 
 using namespace searchModule;
